Switch to new method of playback
Every time a collection is played the user's last playback playlist is removed and a new one is
created. The memory of the last playback playlist is accomplished through a new user db column.

There are many ways that seem like they should work to accomplish this task but many of them have
issues of sporadic failure. Using a newly created playlist seems like a good way to guarantee that
playback should succeed and not play the wrong thing.